requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.4
